Add prometheus support
Report metrics for FIO tests directly to a prometheus push gateway
when the fio action is run.

Push gateway is configured using a configuration option.

Results are reported at the end of each test and don't report
performance over the time the test takes to complete.

Add runtime option to fio action to allow fixed duration tests to
be executed.