Generate unique RBD name per unit
Ensure that every unit uses an RBD with a unique name, avoiding
focus of horizontal test scale on a single set of blocks which
does not represent how real world load might happen.

Bump default image size to 20G to be more presentative of a real
RBD workload and allow FIO tests to spread better across the block
device.