Generate swift key early
The transactional nature of ops framework means that a failed test would
not commit the generated swift key. Thus, the next hook execution the user
would already exist, and a new (now incorrect) key would be generated.

Avoid this by generating the key, storing and setting it on the peer
relation early in the charm's life cycle.

Remove previous attempts at solving this with self._store.