Use datetime to detect end of test runtime
Sometimes fio will take longer than the configured test duration
to complete. The existing code made the assumption that it always
returned in the configured duration so switch to using a time
based calculation which should result in more accurate test
runtimes especially for larger block sizes with high concurrency.