Zero disk devices before benchmarks
For disk devices provided by disk-devices to the fio action,
or devices provided by the `test-devices` storage with juju.

This ensures that the disk is fully initialised,
so any kind of thin-provisioning won't affect the benchmarks.

Also update the description with a warning.
Note that data would have previously been lost anyway by the fio test,
not just from zeroing the disk.
But either way, we should make it clear in the docs.

Closes-Bug: https://bugs.launchpad.net/charm-woodpecker/+bug/2043167