Add unit tests for woodpecker charm
This PR adds unit tests so that changes can be more easily checked
for consistency and correctness. In addition, it also makes the
swift-bench optional instead of mandatory, given that it's not
very useful (IMO), and it requires a binary snap file to be provided,
which is not very ergonomic nor user friendly.

When it comes to requirements, this PR aligns itself with the Jammy
enablement PR so that it can be integrated more easily.