Revert "Make template rendering take operation type into account"
This reverts commit 01d4c9a2675d31b5a7afeb483dd65451636405eb.

Silently ignoring action_params.iodepth that is explicitly passed by an
user without any feedback is not intuitive or user friendly. When
somebody runs the action with block-size=4k iodepth=32 and
operation=randwrite or randread, randwrite gives a better number than
randread, which is surprising. But it's just because iodepth=1 was
hardcoded for randread. That doesn't help to generate data for
comparison.

Emptying the default value for iodepth by having a logic of recommendded
values or allowing specifying "time_based" can be implemented separately
without hard coding.

Closes-Bug: #2057461
Related-Bug: #1940371
Related-Bug: #2043167